requirements.txt
fastapi==0.103.0
uvicorn==0.23.2
pyyaml
rich==13.5.2
func_timeout==4.3.5
